move error into separate module
